# perf: save a single snapshot of warning filters and use it